Created the endpoint to read from csv, need to organize the code following clean architecture
