+++ app/files/characters.csv
@@ -1,0 +1,5 @@
+﻿1,Rick Sanchez
+2,Morty Smith
+3,Summer Smith
+4,Beth Smith
+5,Jerry Smith
